+++ GRDB/FTS/FTS3.swift
@@ -56,7 +56,7 @@ public final class FTS3TableDefinition {
     ///         t.tokenizer = .porter
     ///     }
     /// See https://www.sqlite.org/fts3.html#creating_and_destroying_fts_tables
-    public var tokenizer: FTS3TokenizerRequest?
+    public var tokenizer: FTS3TokenizerDescriptor?
     
     /// Appends a table column.
     ///