+++ Tests/Public/FTS/FTS5CustomTokenizerTests.swift
@@ -18,7 +18,7 @@ private final class StopWordsTokenizer : FTS5CustomTokenizer {
         if arguments.isEmpty {
             wrappedTokenizer = try db.makeTokenizer(.unicode61())
         } else {
-            wrappedTokenizer = try db.makeTokenizer(FTS5TokenizerRequest(components: arguments))
+            wrappedTokenizer = try db.makeTokenizer(FTS5TokenizerDescriptor(components: arguments))
         }
         // TODO: find a way to provide stop words through arguments
         ignoredTokens = ["bar"]
@@ -73,7 +73,7 @@ private final class NFKCTokenizer : FTS5CustomTokenizer {
         if arguments.isEmpty {
             wrappedTokenizer = try db.makeTokenizer(.unicode61())
         } else {
-            wrappedTokenizer = try db.makeTokenizer(FTS5TokenizerRequest(components: arguments))
+            wrappedTokenizer = try db.makeTokenizer(FTS5TokenizerDescriptor(components: arguments))
         }
     }
     
@@ -131,7 +131,7 @@ private final class SynonymsTokenizer : FTS5CustomTokenizer {
         if arguments.isEmpty {
             wrappedTokenizer = try db.makeTokenizer(.unicode61())
         } else {
-            wrappedTokenizer = try db.makeTokenizer(FTS5TokenizerRequest(components: arguments))
+            wrappedTokenizer = try db.makeTokenizer(FTS5TokenizerDescriptor(components: arguments))
         }
         synonyms = [["first", "1st"]]
     }
@@ -207,7 +207,7 @@ class FTS5CustomTokenizerTests: GRDBTestCase {
             try dbQueue.inDatabase { db in
                 try db.create(virtualTable: "documents", using: FTS5()) { t in
                     // TODO: improve this API
-                    t.tokenizer = FTS5TokenizerRequest(name: StopWordsTokenizer.name)
+                    t.tokenizer = FTS5TokenizerDescriptor(name: StopWordsTokenizer.name)
                     t.column("content")
                 }
                 
@@ -231,7 +231,7 @@ class FTS5CustomTokenizerTests: GRDBTestCase {
             
             try dbPool.write { db in
                 try db.create(virtualTable: "documents", using: FTS5()) { t in
-                    t.tokenizer = FTS5TokenizerRequest(name: StopWordsTokenizer.name)
+                    t.tokenizer = FTS5TokenizerDescriptor(name: StopWordsTokenizer.name)
                     t.column("content")
                 }
                 
@@ -281,7 +281,7 @@ class FTS5CustomTokenizerTests: GRDBTestCase {
             // With NFKC conversion wrapping unicode61 (the default)
             try dbQueue.inDatabase { db in
                 try db.create(virtualTable: "documents", using: FTS5()) { t in
-                    t.tokenizer = FTS5TokenizerRequest(name: NFKCTokenizer.name)
+                    t.tokenizer = FTS5TokenizerDescriptor(name: NFKCTokenizer.name)
                     t.column("content")
                 }
                 
@@ -297,8 +297,8 @@ class FTS5CustomTokenizerTests: GRDBTestCase {
             // With NFKC conversion wrapping ascii
             try dbQueue.inDatabase { db in
                 try db.create(virtualTable: "documents", using: FTS5()) { t in
-                    let ascii = FTS5TokenizerRequest.ascii()
-                    t.tokenizer = FTS5TokenizerRequest(name: NFKCTokenizer.name, arguments: ascii.components)
+                    let ascii = FTS5TokenizerDescriptor.ascii()
+                    t.tokenizer = FTS5TokenizerDescriptor(name: NFKCTokenizer.name, arguments: ascii.components)
                     t.column("content")
                 }
                 
@@ -320,7 +320,7 @@ class FTS5CustomTokenizerTests: GRDBTestCase {
             
             try dbQueue.inDatabase { db in
                 try db.create(virtualTable: "documents", using: FTS5()) { t in
-                    t.tokenizer = FTS5TokenizerRequest(name: SynonymsTokenizer.name)
+                    t.tokenizer = FTS5TokenizerDescriptor(name: SynonymsTokenizer.name)
                     t.column("content")
                 }